+++ conanfile.py
@@ -3,13 +3,13 @@ from conan.tools.meson import MesonToolchain, Meson
 
 class VerySimpleSMTPSConan(ConanFile):
     generators = "PkgConfigDeps"
-    settings = "os", "arch", "compiler", "build_type"
+    settings = "os", "compiler", "build_type"
     options = {"shared": [True, False]}
     default_options = {"shared": False}
 
     required_packages = [
         "libcurl/8.0.1",
-        "cpputest/4.0"
+        "cpputest/4.0",
     ]
 
     def requirements(self):
