+++ src/utils/secure_strings.hpp
